R7.2: split vector/assemble into model/materialize/check
Facade mod.rs re-exports everything; no public path changes. Producer
materialization (assemble_vector_fir, state storage/loop/island/top
level materializers and their contexts) lives in materialize.rs; the
independent checker in check.rs; artifact DTOs, the verified wrapper,
the error taxonomy and shared total conversions
(scheduled_island_loop_ids, recursion_name) in model.rs.

Plan §4.8 — admission guards in the shared verify function
`verify_vector_fir_assembly` (hosted in check.rs):
- require_same_plan / cursor equations
  (independently_expected_clock_cursor, state_cursor_advance_matches)
- verify_assembled_fused_serial_groups / verify_assembled_lockstep_bundles
- verify_clock_output_stores / verify_action_shape / guard_shape_matches
- expected_island_declarations

remain on BOTH paths after the split: the producer's terminal step
(assemble_vector_fir) and standalone checker callers. Existing
rejection tests fail through the checker entry alone (forged
assemblies) and stay green. Producer/checker counterparts carry
module-header merge prohibitions (plan §3.2).

Validation: transform 387 lib tests, transform clippy (+1.97.0) clean,
release compiler build clean (lib without cfg(test) verified).
