+++ porting/journal/2026-07-20.md
@@ -133,3 +133,22 @@ modules on both sides of the plan artifact. Validation: transform 385,
   release de l'arbitre — échouait. Détecté par l'échec de l'arbitre,
   gates re-validées après retrait des gates bidon ; leçon : valider
   `cargo build -p transform` (lib sans cfg test) après chaque cargo fix.
+
+## R7.2 — scission de `vector/assemble`
+
+- `assemble/mod.rs` (2 684 lignes) scindé en `model.rs` (249),
+  `materialize.rs` (1 346), `check.rs` (1 054), façade `mod.rs` (41).
+- §4.8 : `verify_vector_fir_assembly` (check.rs) reste appelé par l'étape
+  terminale du producteur (`assemble_vector_fir`) et par les appels
+  autonomes (module, tests). Tests de rejet par l'entrée checker seule
+  déjà présents (assemblages forgés) — inchangés et verts.
+- Duplication d'assurance documentée dans les deux en-têtes :
+  matérialisations producteur vs re-dérivations checker
+  (independently_expected_clock_cursor, state_cursor_advance_matches,
+  expected_island_declarations, shape matchers).
+- Partagés (§4.6, conversions totales) dans model.rs :
+  scheduled_island_loop_ids, recursion_name.
+- `fir_reachable` élargi à pub(in …::vector) (importé par module).
+- Validation lib SANS cfg(test) incluse cette fois (build release
+  compiler) après le leçon de R7.1.
+- Gates : transform 387, clippy 0, fmt vert, release compiler vert.
